+++ sdks/python/hatchet_sdk/__init__.py
@@ -76,7 +76,6 @@ from hatchet_sdk.clients.rest.models.pull_request_state import PullRequestState
 from hatchet_sdk.clients.rest.models.reject_invite_request import RejectInviteRequest
 from hatchet_sdk.clients.rest.models.replay_event_request import ReplayEventRequest
 from hatchet_sdk.clients.rest.models.rerun_step_run_request import RerunStepRunRequest
-from hatchet_sdk.clients.rest.models.step import Step
 from hatchet_sdk.clients.rest.models.step_run import StepRun
 from hatchet_sdk.clients.rest.models.step_run_diff import StepRunDiff
 from hatchet_sdk.clients.rest.models.step_run_status import StepRunStatus
@@ -138,15 +137,16 @@ from hatchet_sdk.contracts.workflows_pb2 import (
     WorkerLabelComparator,
 )
 from hatchet_sdk.hatchet import Hatchet
-from hatchet_sdk.utils.aio_utils import sync_to_async
-from hatchet_sdk.worker.worker import Worker, WorkerStartOptions, WorkerStatus
-from hatchet_sdk.workflow import (
-    BaseWorkflow,
+from hatchet_sdk.runnables.task import Task
+from hatchet_sdk.runnables.types import (
     ConcurrencyExpression,
     ConcurrencyLimitStrategy,
+    EmptyModel,
     StickyStrategy,
     WorkflowConfig,
 )
+from hatchet_sdk.utils.aio_utils import sync_to_async
+from hatchet_sdk.worker.worker import Worker, WorkerStartOptions, WorkerStatus
 
 __all__ = [
     "AcceptInviteRequest",
@@ -191,7 +191,6 @@ __all__ = [
     "RejectInviteRequest",
     "ReplayEventRequest",
     "RerunStepRunRequest",
-    "Step",
     "StepRun",
     "StepRunDiff",
     "StepRunStatus",
@@ -243,14 +242,13 @@ __all__ = [
     "WorkerContext",
     "ClientConfig",
     "Hatchet",
-    "concurrency",
-    "on_failure_step",
-    "step",
     "workflow",
     "Worker",
     "WorkerStartOptions",
     "WorkerStatus",
     "ConcurrencyExpression",
-    "BaseWorkflow",
+    "Workflow",
     "WorkflowConfig",
+    "Task",
+    "EmptyModel",
 ]
