+++ sdks/python/hatchet_sdk/worker/runner/run_loop_manager.py
@@ -8,11 +8,11 @@ from hatchet_sdk.client import Client, new_client_raw
 from hatchet_sdk.clients.dispatcher.action_listener import Action
 from hatchet_sdk.config import ClientConfig
 from hatchet_sdk.logger import logger
+from hatchet_sdk.runnables.task import Task
 from hatchet_sdk.utils.typing import WorkflowValidator
 from hatchet_sdk.worker.action_listener_process import ActionEvent
 from hatchet_sdk.worker.runner.runner import Runner
 from hatchet_sdk.worker.runner.utils.capture_logs import capture_logs
-from hatchet_sdk.workflow import Step
 
 STOP_LOOP_TYPE = Literal["STOP_LOOP"]
 STOP_LOOP: STOP_LOOP_TYPE = "STOP_LOOP"
@@ -23,7 +23,7 @@ T = TypeVar("T")
 @dataclass
 class WorkerActionRunLoopManager:
     name: str
-    action_registry: dict[str, Step[Any]]
+    action_registry: dict[str, Task[Any, Any]]
     validator_registry: dict[str, WorkflowValidator]
     max_runs: int | None
     config: ClientConfig
