+++ sdks/python/hatchet_sdk/worker/runner/runner.py
@@ -8,7 +8,7 @@ from concurrent.futures import ThreadPoolExecutor
 from enum import Enum
 from multiprocessing import Queue
 from threading import Thread, current_thread
-from typing import Any, Callable, Dict, TypeVar, cast
+from typing import Any, Callable, Dict, cast
 
 from pydantic import BaseModel
 
@@ -30,12 +30,11 @@ from hatchet_sdk.contracts.dispatcher_pb2 import (
     STEP_EVENT_TYPE_STARTED,
 )
 from hatchet_sdk.logger import logger
+from hatchet_sdk.runnables.task import Task
+from hatchet_sdk.runnables.types import R, TWorkflowInput
 from hatchet_sdk.utils.typing import WorkflowValidator
 from hatchet_sdk.worker.action_listener_process import ActionEvent
 from hatchet_sdk.worker.runner.utils.capture_logs import copy_context_vars, sr, wr
-from hatchet_sdk.workflow import Step
-
-T = TypeVar("T")
 
 
 class WorkerStatus(Enum):
@@ -52,7 +51,7 @@ class Runner:
         event_queue: "Queue[Any]",
         max_runs: int | None = None,
         handle_kill: bool = True,
-        action_registry: dict[str, Step[T]] = {},
+        action_registry: dict[str, Task[TWorkflowInput, R]] = {},
         validator_registry: dict[str, WorkflowValidator] = {},
         config: ClientConfig = ClientConfig(),
         labels: dict[str, str | int] = {},
@@ -64,7 +63,7 @@ class Runner:
         self.max_runs = max_runs
         self.tasks: dict[str, asyncio.Task[Any]] = {}  # Store run ids and futures
         self.contexts: dict[str, Context] = {}  # Store run ids and contexts
-        self.action_registry: dict[str, Step[T]] = action_registry
+        self.action_registry: dict[str, Task[TWorkflowInput, R]] = action_registry
         self.validator_registry = validator_registry
 
         self.event_queue = event_queue
@@ -195,7 +194,9 @@ class Runner:
 
         return inner_callback
 
-    def thread_action_func(self, context: Context, step: Step[T], action: Action) -> T:
+    def thread_action_func(
+        self, context: Context, task: Task[TWorkflowInput, R], action: Action
+    ) -> R:
         if action.step_run_id is not None and action.step_run_id != "":
             self.threads[action.step_run_id] = current_thread()
         elif (
@@ -204,22 +205,22 @@ class Runner:
         ):
             self.threads[action.get_group_key_run_id] = current_thread()
 
-        return step.call(context)
+        return task.call(context)
 
     # We wrap all actions in an async func
     async def async_wrapped_action_func(
         self,
         context: Context,
-        step: Step[T],
+        task: Task[TWorkflowInput, R],
         action: Action,
         run_id: str,
-    ) -> T:
+    ) -> R:
         wr.set(context.workflow_run_id)
         sr.set(context.step_run_id)
 
         try:
-            if step.is_async_function:
-                return await step.aio_call(context)
+            if task.is_async_function:
+                return await task.aio_call(context)
             else:
                 pfunc = functools.partial(
                     # we must copy the context vars to the new thread, as only asyncio natively supports
@@ -228,7 +229,7 @@ class Runner:
                     contextvars.copy_context().items(),
                     self.thread_action_func,
                     context,
-                    step,
+                    task,
                     action,
                 )
