+++ sdks/python/hatchet_sdk/worker/worker.py
@@ -21,6 +21,8 @@ from hatchet_sdk.clients.dispatcher.action_listener import Action
 from hatchet_sdk.config import ClientConfig
 from hatchet_sdk.contracts.workflows_pb2 import CreateWorkflowVersionOpts
 from hatchet_sdk.logger import logger
+from hatchet_sdk.runnables.task import Task
+from hatchet_sdk.runnables.workflow import Workflow
 from hatchet_sdk.utils.typing import WorkflowValidator, is_basemodel_subclass
 from hatchet_sdk.worker.action_listener_process import (
     ActionEvent,
@@ -30,10 +32,8 @@ from hatchet_sdk.worker.runner.run_loop_manager import (
     STOP_LOOP_TYPE,
     WorkerActionRunLoopManager,
 )
-from hatchet_sdk.workflow import BaseWorkflow, Step, StepType, Task
 
 T = TypeVar("T")
-TBaseWorkflow = TypeVar("TBaseWorkflow", bound=BaseWorkflow)
 
 
 class WorkerStatus(Enum):
@@ -58,6 +58,7 @@ class Worker:
         debug: bool = False,
         owned_loop: bool = True,
         handle_kill: bool = True,
+        workflows: list[Workflow[Any]] = [],
     ) -> None:
         self.name = name
         self.config = config
@@ -69,7 +70,7 @@ class Worker:
 
         self.client: Client
 
-        self.action_registry: dict[str, Step[Any]] = {}
+        self.action_registry: dict[str, Task[Any, Any]] = {}
         self.validator_registry: dict[str, WorkflowValidator] = {}
 
         self.killing: bool = False
@@ -95,6 +96,8 @@ class Worker:
             "hatchet_worker_status", "Current status of the Hatchet worker"
         )
 
+        self.register_workflows(workflows)
+
     def register_workflow_from_opts(
         self, name: str, opts: CreateWorkflowVersionOpts
     ) -> None:
@@ -105,7 +108,7 @@ class Worker:
             logger.error(e)
             sys.exit(1)
 
-    def register_workflow(self, workflow: TBaseWorkflow) -> None:
+    def register_workflow(self, workflow: Workflow[Any]) -> None:
         namespace = self.client.config.namespace
 
         try:
@@ -117,7 +120,7 @@ class Worker:
             logger.error(e)
             sys.exit(1)
 
-        for step in workflow.steps:
+        for step in workflow.__tasks:
             action_name = workflow.create_action_name(namespace, step)
             self.action_registry[action_name] = step
             return_type = get_type_hints(step.fn).get("return")
@@ -127,31 +130,9 @@ class Worker:
                 step_output=return_type if is_basemodel_subclass(return_type) else None,
             )
 
-    def register_function(self, function: Task[Any, Any]) -> None:
-        from hatchet_sdk.workflow import BaseWorkflow
-
-        declaration = function.hatchet.declare_workflow(
-            **function.workflow_config.model_dump()
-        )
-
-        class Workflow(BaseWorkflow):
-            config = declaration.config
-
-            @property
-            def default_steps(self) -> list[Step[Any]]:
-                return [function.step]
-
-            @property
-            def on_failure_steps(self) -> list[Step[Any]]:
-                if not function.on_failure_step:
-                    return []
-
-                step = function.on_failure_step.step
-                step.type = StepType.ON_FAILURE
-
-                return [step]
-
-        self.register_workflow(Workflow())
+    def register_workflows(self, workflows: list[Workflow[Any]]) -> None:
+        for workflow in workflows:
+            self.register_workflow(workflow)
 
     def status(self) -> WorkerStatus:
         return self._status
