+++ sdks/python/lint.sh
@@ -1,3 +1,6 @@
+echo "\nLinting with ruff"
+poetry run ruff check . --fix
+
 echo "Formatting with black"
 poetry run black . --color
 
@@ -6,6 +9,3 @@ poetry run isort .
 
 echo "\nType checking with mypy"
 poetry run mypy --config-file=pyproject.toml
-
-echo "\nLinting with ruff"
-poetry run ruff check . --fix
